fix: for BCC
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk@166 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/CHokanMgr.cpp
@@ -313,7 +313,7 @@ int CHokanMgr::Search(
 		nX -= nCX - 8;
 	}else{
 		// サイズを調整して右に表示
-		nCX = max(rcDesktop.right - nX , 100);	// 最低サイズを100くらいに
+		nCX = max((int)(rcDesktop.right - nX) , 100);	// 最低サイズを100くらいに
 	}
 
 //	2001/06/19 End
